jurisprudencia-platform/requirements.txt
# Core dependencies
requests>=2.31.0
beautifulsoup4>=4.12.0
PyPDF2>=3.0.0
python-dotenv>=1.0.0

# Database
SQLAlchemy>=2.0.0
psycopg2-binary>=2.9.0
alembic>=1.13.0

# Machine Learning & NLP
sentence-transformers>=2.2.0
torch>=2.0.0
transformers>=4.35.0
numpy>=1.24.0
pandas>=2.0.0

# Vector Database (development)
chromadb>=0.4.0

# API Framework
fastapi>=0.104.0
uvicorn>=0.24.0
pydantic>=2.5.0
python-multipart>=0.0.6

# Authentication
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4

# Task Queue
celery>=5.3.0
redis>=5.0.0
flower>=2.0.0  # Celery monitoring

# Web Interface
streamlit>=1.28.0
plotly>=5.18.0

# Utilities
tqdm>=4.66.0
click>=8.1.0
rich>=13.7.0  # Better terminal output

# Development
pytest>=7.4.0
pytest-asyncio>=0.21.0
black>=23.11.0
flake8>=6.1.0
mypy>=1.7.0
ipython>=8.17.0

# Monitoring
prometheus-client>=0.19.0
opentelemetry-api>=1.21.0
opentelemetry-sdk>=1.21.0

# Documentation
mkdocs>=1.5.0
mkdocs-material>=9.4.0